feat(curriculum): add new challenge, remove outdated
One of the challenges was using very old approaches, so rather than
rewriting it we have removed it entirely. To replace that, we have
created a new challenge that teaches a different concept, but can
be used for the same result. We've also renamed a challenge to
more accurately reflect what it teaches.